keypad: switch to double for division

+++ keypad.c
@@ -47,7 +47,7 @@ int main() {
                     printf("%d * %d = %d\n",num1,num2,num1*num2);
                     break;
                 case 'D': 
-                    printf("%d / %d = %d\n",num1,num2,num1/num2);
+                    printf("%d / %d = %lf\n",num1,num2,(double)num1/num2);
                     break;
                 default:
                     printf("Invalid!\n");
